+++ eqcorrscan/utils/synth_seis.py
@@ -238,7 +238,8 @@ def template_grid(stations, nodes, travel_times, phase, PS_ratio=1.68,
 
 
 def generate_synth_data(nsta, ntemplates, nseeds, samp_rate, t_length,
-                        max_amp, max_lag, phaseout="all", jitter=0):
+                        max_amp, max_lag, phaseout="all", jitter=0,
+                        noise=True, same_phase=False):
     """
     Generate a synthetic dataset to be used for testing.
 
@@ -267,6 +268,10 @@ def generate_synth_data(nsta, ntemplates, nseeds, samp_rate, t_length,
     :type jitter: int
     :param jitter:
         Random range to allow arrival shifts for seeded phases (samples)
+    :type noise: bool
+    :param noise: Set to False to give noise-free data.
+    :type same_phase: bool
+    :param same_phase: Whether to enforce all positive repeats (True) or not.
 
     :returns: Templates: List of :class:`obspy.core.stream.Stream`
     :rtype: list
@@ -275,6 +280,7 @@ def generate_synth_data(nsta, ntemplates, nseeds, samp_rate, t_length,
     :returns: Seeds: dictionary of seed SNR and time with time in samples.
     :rtype: dict
     """
+    jitter = abs(jitter)
     # Generate random arrival times
     t_times = np.abs(np.random.random([nsta, ntemplates])) * max_lag
     # Generate random node locations - these do not matter as they are only
@@ -313,6 +319,8 @@ def generate_synth_data(nsta, ntemplates, nseeds, samp_rate, t_length,
         impulse_times = np.random.randint(86400 * int(samp_rate),
                                           size=nseeds)
         impulse_amplitudes = np.random.randn(nseeds) * max_amp
+        if same_phase:
+            impulse_amplitudes = np.abs(impulse_amplitudes)
         # Generate amplitudes up to maximum amplitude in a normal distribution
         seeds.append({'SNR': impulse_amplitudes,
                       'time': impulse_times})
@@ -324,7 +332,8 @@ def generate_synth_data(nsta, ntemplates, nseeds, samp_rate, t_length,
                        for template_tr in template])
         for j, template_tr in enumerate(template):
             offset = int((template_tr.stats.starttime - mintime) * samp_rate)
-            offset += np.random.randint(-jitter, jitter)
+            if jitter > 0:
+                offset += np.random.randint(-jitter, jitter)
             pad = np.zeros(abs(offset))
             if offset > 0:
                 tr_impulses = np.append(pad, impulses)[0:len(impulses)]
@@ -334,9 +343,10 @@ def generate_synth_data(nsta, ntemplates, nseeds, samp_rate, t_length,
             data[j].data += np.convolve(tr_impulses,
                                         template_tr.data)[0:len(impulses)]
     # Add the noise
-    for tr in data:
-        noise = np.random.randn(86400 * int(samp_rate))
-        tr.data += noise / max(noise)
+    if noise:
+        for tr in data:
+            noise_array = np.random.randn(86400 * int(samp_rate))
+            tr.data += noise_array / max(noise_array)
     return templates, data, seeds