+++ CHANGES.md
@@ -5,6 +5,10 @@
     are faster and more memory efficient.
   - Party.lag_calc and Family.lag_calc now work in-place on the events in 
     the grouping.
+  - Added relative_mags method to Party and Family; this can be called from
+    lag-calc to avoid reprocessing data.
+  - Added lag_calc.xcorr_pick_family as a public facing API to implement
+    correlation re-picking of a group of events.
 * Renamed utils.clustering.cross_chan_coherence to 
   utils.clustering.cross_chan_correlation to better reflect what it actually 
   does.